requirements.txt
fastapi
uvicorn[standard]
playwright
PyMuPDF
beautifulsoup4
sqlalchemy
pydantic
python-dotenv
openai
redis
httpx
pyyaml
python-multipart
aiofiles
# Optional: encryption
cryptography
# Optional: testing + template utilities
pytest
jinja2
fakeredis